requirements.txt
django
djangorestframework
RPi.GPIO
pytz
uwsgi